use FennecPrecompile & refactored
