+++ mix.exs
@@ -12,7 +12,7 @@ defmodule ElixirMake.Mixfile do
       package: package(),
       docs: docs(),
       deps: deps(),
-      make_precompiler: FennecPrecompile.Precompiler
+      make_precompiler: FennecPrecompile
     ]
   end
